requirements.txt
pandas
numpy
azure-storage-blob
azure-identity
pyodbc 
cryptography